Renames syncSlice to idbSlice
