+++ react-front-end/src/app/store.ts
@@ -3,14 +3,14 @@ import { Action, configureStore, ThunkAction } from "@reduxjs/toolkit";
 import { baseApiSlice } from "@/services/baseApiSlice";
 
 import authReducer from "@/features/auth/authSlice";
-import syncReducer from "@/features/offline/syncSlice";
+import idbReducer from "@/features/offline/idbSlice";
 import authListenerMiddleware from "../features/auth/authListenerMiddleware";
 
 export const store = configureStore({
     reducer: {
         [baseApiSlice.reducerPath]: baseApiSlice.reducer,
         auth: authReducer,
-        sync: syncReducer
+        idb: idbReducer
     },
     middleware: getDefaultMiddleware =>
         getDefaultMiddleware().concat(baseApiSlice.middleware, authListenerMiddleware.middleware)
